+++ zcash_client_sqlite/src/wallet/scanning.rs
@@ -581,9 +581,8 @@ pub(crate) fn update_chain_tip<P: consensus::Parameters>(
 pub(crate) mod tests {
     use std::num::NonZeroU8;
 
-    use incrementalmerkletree::{frontier::Frontier, Hashable, Position};
+    use incrementalmerkletree::{frontier::Frontier, Position};
 
-    use sapling::Node;
     use secrecy::SecretVec;
     use zcash_client_backend::data_api::{
         chain::{ChainState, CommitmentTreeRoot},
@@ -599,7 +598,10 @@ pub(crate) mod tests {
 
     use crate::{
         error::SqliteClientError,
-        testing::{pool::ShieldedPoolTester, AddressType, BlockCache, TestBuilder, TestState},
+        testing::{
+            pool::ShieldedPoolTester, AddressType, BlockCache, InitialChainState, TestBuilder,
+            TestState,
+        },
         wallet::{
             sapling::tests::SaplingPoolTester,
             scanning::{insert_queue_entries, replace_queue_entries, suggest_scan_ranges},
@@ -609,7 +611,7 @@ pub(crate) mod tests {
 
     #[cfg(feature = "orchard")]
     use {
-        crate::wallet::orchard::tests::OrchardPoolTester, orchard::tree::MerkleHashOrchard,
+        crate::wallet::orchard::tests::OrchardPoolTester,
         zcash_client_backend::data_api::ORCHARD_SHARD_HEIGHT,
     };
 
@@ -626,66 +628,75 @@ pub(crate) mod tests {
 
     fn scan_complete<T: ShieldedPoolTester>() {
         use ScanPriority::*;
-        let initial_height_offset = 310;
-        let mut st = TestBuilder::new()
-            .with_block_cache()
-            .with_test_account(AccountBirthday::from_sapling_activation)
-            .build();
-
-        let sapling_activation_height = st.sapling_activation_height();
 
         // We'll start inserting leaf notes 5 notes after the end of the third subtree, with a gap
         // of 10 blocks. After `scan_cached_blocks`, the scan queue should have a requested scan
         // range of 300..310 with `FoundNote` priority, 310..320 with `Scanned` priority.
         // We set both Sapling and Orchard to the same initial tree size for simplicity.
+        let prior_block_hash = BlockHash([0; 32]);
         let initial_sapling_tree_size: u32 = (0x1 << 16) * 3 + 5;
         let initial_orchard_tree_size: u32 = (0x1 << 16) * 3 + 5;
+        let initial_height_offset = 310;
 
-        // Construct a fake chain state for the end of block 300
-        let (prior_sapling_frontiers, sapling_initial_tree) =
-            Frontier::random_with_prior_subtree_roots(
-                st.rng(),
-                initial_sapling_tree_size.into(),
-                NonZeroU8::new(16).unwrap(),
-            );
-        let sapling_subtree_roots = prior_sapling_frontiers
-            .into_iter()
-            .zip(0u32..)
-            .map(|(root, i)| {
-                CommitmentTreeRoot::from_parts(sapling_activation_height + (100 * (i + 1)), root)
-            })
-            .collect::<Vec<_>>();
+        let mut st = TestBuilder::new()
+            .with_block_cache()
+            .with_initial_chain_state(|rng, network| {
+                let sapling_activation_height =
+                    network.activation_height(NetworkUpgrade::Sapling).unwrap();
+                // Construct a fake chain state for the end of block 300
+                let (prior_sapling_roots, sapling_initial_tree) =
+                    Frontier::random_with_prior_subtree_roots(
+                        rng,
+                        initial_sapling_tree_size.into(),
+                        NonZeroU8::new(16).unwrap(),
+                    );
+                let prior_sapling_roots = prior_sapling_roots
+                    .into_iter()
+                    .zip(0u32..)
+                    .map(|(root, i)| {
+                        CommitmentTreeRoot::from_parts(
+                            sapling_activation_height + (100 * (i + 1)),
+                            root,
+                        )
+                    })
+                    .collect::<Vec<_>>();
 
-        #[cfg(feature = "orchard")]
-        let (prior_orchard_frontiers, orchard_initial_tree) =
-            Frontier::random_with_prior_subtree_roots(
-                st.rng(),
-                initial_orchard_tree_size.into(),
-                NonZeroU8::new(16).unwrap(),
-            );
-        #[cfg(feature = "orchard")]
-        let orchard_subtree_roots = prior_orchard_frontiers
-            .into_iter()
-            .zip(0u32..)
-            .map(|(root, i)| {
-                CommitmentTreeRoot::from_parts(sapling_activation_height + (100 * (i + 1)), root)
-            })
-            .collect::<Vec<_>>();
-
-        let prior_block_hash = BlockHash([0; 32]);
-        st.establish_chain_state(
-            ChainState::new(
-                sapling_activation_height + initial_height_offset - 1,
-                prior_block_hash,
-                sapling_initial_tree,
                 #[cfg(feature = "orchard")]
-                orchard_initial_tree,
-            ),
-            &sapling_subtree_roots,
-            #[cfg(feature = "orchard")]
-            &orchard_subtree_roots,
-        )
-        .unwrap();
+                let (prior_orchard_roots, orchard_initial_tree) =
+                    Frontier::random_with_prior_subtree_roots(
+                        rng,
+                        initial_orchard_tree_size.into(),
+                        NonZeroU8::new(16).unwrap(),
+                    );
+                #[cfg(feature = "orchard")]
+                let prior_orchard_roots = prior_orchard_roots
+                    .into_iter()
+                    .zip(0u32..)
+                    .map(|(root, i)| {
+                        CommitmentTreeRoot::from_parts(
+                            sapling_activation_height + (100 * (i + 1)),
+                            root,
+                        )
+                    })
+                    .collect::<Vec<_>>();
+
+                InitialChainState {
+                    chain_state: ChainState::new(
+                        sapling_activation_height + initial_height_offset - 1,
+                        prior_block_hash,
+                        sapling_initial_tree,
+                        #[cfg(feature = "orchard")]
+                        orchard_initial_tree,
+                    ),
+                    prior_sapling_roots,
+                    #[cfg(feature = "orchard")]
+                    prior_orchard_roots,
+                }
+            })
+            .with_account_from_sapling_activation(BlockHash([3; 32]))
+            .build();
+
+        let sapling_activation_height = st.sapling_activation_height();
 
         let dfvk = T::test_account_fvk(&st);
         let value = NonNegativeAmount::const_from_u64(50000);
@@ -776,37 +787,43 @@ pub(crate) mod tests {
     }
 
     pub(crate) fn test_with_nu5_birthday_offset<T: ShieldedPoolTester>(
-        offset: u32,
+        birthday_offset: u32,
+        prior_block_hash: BlockHash,
     ) -> (TestState<BlockCache>, T::Fvk, AccountBirthday, u32) {
         let st = TestBuilder::new()
             .with_block_cache()
-            .with_test_account(|network| {
+            .with_account_birthday(|rng, network, initial_chain_state| {
+                // We're constructing the birthday without adding any chain data.
+                assert!(initial_chain_state.is_none());
+
                 // We set the Sapling and Orchard frontiers at the birthday height to be
                 // 1234 notes into the second shard.
-                let birthday_height =
-                    network.activation_height(NetworkUpgrade::Nu5).unwrap() + offset;
                 let frontier_position = Position::from((0x1 << 16) + 1234);
-                let sapling_frontier = Frontier::from_parts(
-                    frontier_position,
-                    Node::empty_leaf(),
-                    vec![Node::empty_leaf(); frontier_position.past_ommer_count().into()],
-                )
-                .unwrap();
+                let birthday_height =
+                    network.activation_height(NetworkUpgrade::Nu5).unwrap() + birthday_offset;
+
+                // Construct a fake chain state for the end of the block with the given
+                // birthday_offset from the Nu5 birthday.
+                let (_, sapling_initial_tree) = Frontier::random_with_prior_subtree_roots(
+                    rng,
+                    (frontier_position + 1).into(),
+                    NonZeroU8::new(16).unwrap(),
+                );
                 #[cfg(feature = "orchard")]
-                let orchard_frontier = Frontier::from_parts(
-                    frontier_position,
-                    MerkleHashOrchard::empty_leaf(),
-                    vec![
-                        MerkleHashOrchard::empty_leaf();
-                        frontier_position.past_ommer_count().into()
-                    ],
-                )
-                .unwrap();
+                let (_, orchard_initial_tree) = Frontier::random_with_prior_subtree_roots(
+                    rng,
+                    (frontier_position + 1).into(),
+                    NonZeroU8::new(16).unwrap(),
+                );
+
                 AccountBirthday::from_parts(
-                    birthday_height,
-                    sapling_frontier,
-                    #[cfg(feature = "orchard")]
-                    orchard_frontier,
+                    ChainState::new(
+                        birthday_height,
+                        prior_block_hash,
+                        sapling_initial_tree,
+                        #[cfg(feature = "orchard")]
+                        orchard_initial_tree,
+                    ),
                     None,
                 )
             })
@@ -834,7 +851,8 @@ pub(crate) mod tests {
         use ScanPriority::*;
 
         // Use a non-zero birthday offset because Sapling and NU5 are activated at the same height.
-        let (st, _, birthday, sap_active) = test_with_nu5_birthday_offset::<T>(76);
+        let (st, _, birthday, sap_active) =
+            test_with_nu5_birthday_offset::<T>(76, BlockHash([0; 32]));
         let birthday_height = birthday.height().into();
 
         let expected = vec![
@@ -869,11 +887,8 @@ pub(crate) mod tests {
         st.wallet_mut()
             .create_account(
                 &SecretVec::new(vec![0; 32]),
-                AccountBirthday::from_parts(
-                    wallet_birthday,
-                    Frontier::empty(),
-                    #[cfg(feature = "orchard")]
-                    Frontier::empty(),
+                &AccountBirthday::from_parts(
+                    ChainState::empty(wallet_birthday - 1, BlockHash([0; 32])),
                     None,
                 ),
             )
@@ -904,7 +919,8 @@ pub(crate) mod tests {
         use ScanPriority::*;
 
         // Use a non-zero birthday offset because Sapling and NU5 are activated at the same height.
-        let (mut st, _, birthday, sap_active) = test_with_nu5_birthday_offset::<T>(76);
+        let (mut st, _, birthday, sap_active) =
+            test_with_nu5_birthday_offset::<T>(76, BlockHash([0; 32]));
 
         // Set up the following situation:
         //
@@ -921,7 +937,9 @@ pub(crate) mod tests {
 
         // Verify that the suggested scan ranges match what is expected.
         let expected = vec![
-            // The wallet's birthday onward is marked for recovery.
+            // The wallet's birthday onward is marked for recovery. Because we don't
+            // yet have any chain state, it is marked with `Historic` priority rather
+            // than `ChainTip`.
             scan_range(wallet_birthday..chain_end, Historic),
             // The range below the wallet's birthday height is ignored.
             scan_range(sap_active..wallet_birthday, Ignored),
@@ -946,7 +964,8 @@ pub(crate) mod tests {
         use ScanPriority::*;
 
         // Use a non-zero birthday offset because Sapling and NU5 are activated at the same height.
-        let (mut st, _, birthday, sap_active) = test_with_nu5_birthday_offset::<T>(76);
+        let (mut st, _, birthday, sap_active) =
+            test_with_nu5_birthday_offset::<T>(76, BlockHash([0; 32]));
 
         // Set up the following situation:
         //
@@ -1007,7 +1026,8 @@ pub(crate) mod tests {
 
         // Use a non-zero birthday offset because Sapling and NU5 are activated at the same height.
         // this birthday is 1234 notes into the second shard
-        let (mut st, dfvk, birthday, sap_active) = test_with_nu5_birthday_offset::<T>(76);
+        let (mut st, dfvk, birthday, sap_active) =
+            test_with_nu5_birthday_offset::<T>(76, BlockHash([0; 32]));
 
         // Set up the following situation:
         //
@@ -1149,7 +1169,8 @@ pub(crate) mod tests {
         use ScanPriority::*;
 
         // Use a non-zero birthday offset because Sapling and NU5 are activated at the same height.
-        let (mut st, dfvk, birthday, sap_active) = test_with_nu5_birthday_offset::<T>(76);
+        let (mut st, dfvk, birthday, sap_active) =
+            test_with_nu5_birthday_offset::<T>(76, BlockHash([0; 32]));
 
         // Set up the following situation:
         //